chore: small layout fixes
